avoid duplicate code between two "request-overview"-pages by centralizing a sort-function
