+++ dataland-frontend/src/components/pages/CompanyDataRequestsOverview.vue
@@ -206,6 +206,7 @@ import { getFrontendFrameworkDefinition } from '@/frameworks/FrontendFrameworkRe
 import AuthenticationWrapper from '@/components/wrapper/AuthenticationWrapper.vue';
 import { accessStatusBadgeClass, badgeClass } from '@/utils/RequestUtils';
 import PrimeButton from 'primevue/button';
+import { customCompareForRequestStatus } from '@/utils/RequestsOverviewPageUtils';
 
 export default defineComponent({
   name: 'MyDataRequestsOverview',
@@ -423,7 +424,7 @@ export default defineComponent({
       }
 
       if (a.requestStatus != b.requestStatus)
-        return this.customCompareForRequestStatus(a.requestStatus, b.requestStatus);
+        return customCompareForRequestStatus(a.requestStatus, b.requestStatus, this.sortOrder);
 
       if (a.lastModifiedDate < b.lastModifiedDate) return this.sortOrder;
       if (a.lastModifiedDate > b.lastModifiedDate) return -1 * this.sortOrder;
@@ -431,22 +432,7 @@ export default defineComponent({
       if ((a.userEmailAddress ?? '') < (b.userEmailAddress ?? '')) return -1 * this.sortOrder;
       else return this.sortOrder;
     },
-    /**
-     * Compares two request status
-     * @param a RequestStatus to compare
-     * @param b RequestStatus to compare
-     * @returns result of the comparison
-     */
-    customCompareForRequestStatus(a: RequestStatus, b: RequestStatus) {
-      const sortOrderRequestStatus: { [key: string]: number } = {};
-      sortOrderRequestStatus[RequestStatus.Answered] = 1;
-      sortOrderRequestStatus[RequestStatus.Open] = 2;
-      sortOrderRequestStatus[RequestStatus.Resolved] = 3;
-      sortOrderRequestStatus[RequestStatus.Closed] = 4;
-      sortOrderRequestStatus[RequestStatus.Withdrawn] = 5;
-      if (sortOrderRequestStatus[a] <= sortOrderRequestStatus[b]) return -1 * this.sortOrder;
-      return this.sortOrder;
-    },
+
     /**
      * Updates the access status
      * @param requestId to update
